Added lightnet and created integration tests
