+++ packages/sequencer/package.json
@@ -11,6 +11,7 @@
     "test:file": "node --experimental-vm-modules --experimental-wasm-modules --experimental-wasm-threads ../../node_modules/jest/bin/jest.js",
     "test": "npm run test:file -- ./test/**",
     "test:watch": "npm run test:file -- ./test/** --watch",
+    "integration": "npm run test:file -- ./test-integration/** --runInBand",
     "start": "npm run build && node --experimental-vm-modules --experimental-wasm-modules --experimental-wasm-threads --es-module-specifier-resolution=node ./dist/src/entry.js"
   },
   "main": "dist/index.js",
