+++ packages/sequencer/test-integration/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./../../../tsconfig.json",
+  "compilerOptions": {
+    "experimentalDecorators": true
+  },
+  "include": ["./**/*.ts", "./*.ts"]
+}
